+++ src/basic/src/components/select-option/index.js
@@ -1,7 +1,1 @@
-export const createSelectOption = ({ id, text, disabled }) => {
-	const option = document.createElement('option');
-	option.value = id;
-	option.textContent = text;
-	option.disabled = disabled;
-	return option;
-};
+export { createSelectOption } from './SelectOption';
